drivers: stm32_gpio: add power management
Adds the suspend and resume support.

At resume, the driver restores all the pinctrl configuration that were
applied during the runtime. These configuration are saved in the 'backups'
struct member of the gpio banks.
Hence, the client drivers do not have to apply the pinctrl configuration
at resume, since this is automatically done by the gpio driver

Signed-off-by: Fabien Dessenne <fabien.dessenne@foss.st.com>
Signed-off-by: Gatien Chevallier <gatien.chevallier@st.com>
Change-Id: I23a67872d905b652c1c4212c402a82df1d27ff84