plat-stm32mp1: remove set_etzpc_secure_configuration()
This functionality moved to the firewall driver and should be configured
by device tree instead of statically in an initialisation function. The
previous commit breaks compilation because etzpc_configure_decprot and
etzpc_lock_decprot are no longer exported.

Note that since the device trees haven't been updated yet, this breaks
the ETZPC configuration completely. It will be fixed in a later commit
that introduces the configuration in device tree.

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>